requirements.txt
streamlit
google-cloud-aiplatform>=1.56.0,<2.0.0
google-auth
requests
vertexai